requirements.txt.txt
aiogram
requests
python-dotenv
google-generativeai
apscheduler